task3/task3/my_fastapi_app/requirements.txt
fastapi
uvicorn[standard]
jinja2
passlib[bcrypt]
python-multipart
sqlalchemy
pymysql
python-dotenv
authlib
email-validator
pymysql
